+++ rest/insights/v1/model_insights_v1_metadata_response_cubes_dimensions.go
@@ -17,7 +17,6 @@ package openapi
 // InsightsV1MetadataResponseCubesDimensions struct for InsightsV1MetadataResponseCubesDimensions
 type InsightsV1MetadataResponseCubesDimensions struct {
 	Name        string `json:"name"`
-	DisplayName string `json:"displayName,omitempty"`
 	Description string `json:"description,omitempty"`
 	Type        string `json:"type"`
 }