+++ rest/insights/v1/model_insights_v1_metadata_response_cubes_measures.go
@@ -17,7 +17,6 @@ package openapi
 // InsightsV1MetadataResponseCubesMeasures struct for InsightsV1MetadataResponseCubesMeasures
 type InsightsV1MetadataResponseCubesMeasures struct {
 	Name        string `json:"name"`
-	DisplayName string `json:"displayName,omitempty"`
 	Description string `json:"description,omitempty"`
 	Type        string `json:"type"`
 	Aggregation string `json:"aggregation,omitempty"`