+++ twilio.go
@@ -32,7 +32,6 @@ import (
 	FrontlineV1 "github.com/twilio/twilio-go/rest/frontline/v1"
 	IamScim "github.com/twilio/twilio-go/rest/iam/scim"
 	IamV1 "github.com/twilio/twilio-go/rest/iam/v1"
-	IamV2 "github.com/twilio/twilio-go/rest/iam/v2"
 	InsightsV1 "github.com/twilio/twilio-go/rest/insights/v1"
 	IntelligenceV2 "github.com/twilio/twilio-go/rest/intelligence/v2"
 	IpMessagingV1 "github.com/twilio/twilio-go/rest/ip_messaging/v1"
@@ -93,7 +92,6 @@ type RestClient struct {
 	PreviewIamOrganization *PreviewIamOrganization.ApiService
 	IamScim                *IamScim.ApiService
 	IamV1                  *IamV1.ApiService
-	IamV2                  *IamV2.ApiService
 	InsightsV1             *InsightsV1.ApiService
 	IntelligenceV2         *IntelligenceV2.ApiService
 	IpMessagingV1          *IpMessagingV1.ApiService
@@ -230,7 +228,6 @@ func NewRestClientWithParams(params ClientParams) *RestClient {
 	c.PreviewIamOrganization = PreviewIamOrganization.NewApiService(c.RequestHandler)
 	c.IamScim = IamScim.NewApiService(c.RequestHandler)
 	c.IamV1 = IamV1.NewApiService(c.RequestHandler)
-	c.IamV2 = IamV2.NewApiService(c.RequestHandler)
 	c.InsightsV1 = InsightsV1.NewApiService(c.RequestHandler)
 	c.IntelligenceV2 = IntelligenceV2.NewApiService(c.RequestHandler)
 	c.IpMessagingV1 = IpMessagingV1.NewApiService(c.RequestHandler)